==> 2_code/0_download_xbrl_instances.ipynb <==
import os, re, glob, requests, gzip, urllib.request, time, zipfile, shutil
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm_notebook as tqdm
from bs4 import BeautifulSoup

os.chdir('..')

---

pd.options.display.float_format = '{:.4f}'.format

headers = {"User-Agent": "jo boulement jo@gmx.at"}

directory = '/Users/maltemax/Desktop/xbrl_instances'
external = '/Volumes/T7 Shield/research_data/us_xbrl_instances'

---

df = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'va_data_raw.tsv'),
                 sep = '\t')

df = df.drop_duplicates(subset = 'adsh')

files_desktop = [re.search('(.*?).xml', f)[1] for f in os.listdir(directory) if f.endswith('xml')]

files_external = [re.search('(.*?).xml', f)[1] for f in os.listdir(external) if f.endswith('xml')]

files = files_external + files_desktop 

df = df.loc[~df['adsh'].isin(files)]

for i, r in tqdm(df.iterrows(), total = len(df)):
    content = requests.get(r['url_instance'], headers = headers).content
    with open(os.path.join(directory, f'{r["adsh"]}.xml'), 'wb') as file:
        file.write(content)

---

for file in tqdm([f for f in os.listdir('/Users/maltemax/Desktop/xbrl_instances') if f.endswith('xml')]):
    shutil.move(f"/Users/maltemax/Desktop/xbrl_instances/{file}",
                f"/Volumes/T7 Shield/research_data/us_xbrl_instances/{file}")
