==> 2_code/0_get_compustat_crsp_link.ipynb <==
import os, re, wrds
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm_notebook as tqdm

os.chdir('..')

---

db = wrds.Connection()

ccm = db.get_table(library = 'crsp_a_ccm', table = 'ccmxpf_lnkhist')

following: https://iangow.github.io/far_book/identifiers.html

"One thing you will see is that there are cases where lpermno is NA, so “matching” these rows will result in non-matches, which is of no real value. The only value might be in determining whether the non-match has linktype of NR, which means that lack of a link has been “confirmed by research” (presumably by CRSP), or of NU, which means the link is “not yet confirmed” by research."

ccm = ccm.loc[ccm['lpermno'].notna()]

"The cases where linktype is LD represent cases where two GVKEYs map to a single PERMNO at the same time and, according to WRDS, “this link should not be used.” Here is one example:"

ccm = ccm.loc[ccm['linktype'] != 'LD']

"The cases where linktype is LX represent cases where the security referred to on Compustat is one that trades on a foreign exchange and CRSP is merely “helpfully” linking to a different security that *is* found on CRSP. "

ccm = ccm.loc[ccm['linktype'] != "LX"]

"The remaining category for discussion is where linktype is LN. These are cases where a link exists, but Compustat does not have price data to allow CRSP to check the quality of the link. While researcher discretion might be used to include these, most researchers appear to exclude these cases and we will do likewise. Given the above, we are only including cases where linktype is in LC (valid, researched link), LU (unresearched link), or LS (link valid for this lpermno only)."

ccm = ccm.loc[ccm['linktype'] != "LN"]

"Now, let’s consider, linkprim, which WRDS explains as follows:

`linkprim clarifies the link’s relationship to Compustat’s marked primary security within the related range. “P” indicates a primary link marker, as identified by Compustat in monthly security data. “C” indicates a primary link marker, as identified by CRSP to resolve ranges of overlapping or missing primary markers from Compustat in order to produce one primary security throughout the company history. “J” indicates a joiner secondary issue of a company, identified by Compustat in monthly security data.`


This suggests we should omit cases where linkprim equals J. Given that cases where linkprim equals N are duplicated links due to the existence of Canadian securities for a US-traded firm, we will exclude these too."

ccm = ccm.loc[ccm['linkprim'].isin(['C', 'P'])]

from here on, own code:

ccm['date_from'], ccm['date_to'] = pd.to_datetime(ccm['linkdt']), pd.to_datetime(ccm['linkenddt'])

ccm['fyear_from'] = np.where(ccm['date_from'].dt.month <= 6,
                            ccm['date_from'].dt.year,
                            ccm['date_from'].dt.year + 1)

ccm['fyear_to'] = np.where(ccm['date_to'].dt.month <= 6,
                          ccm['date_to'].dt.year - 1,
                          ccm['date_to'].dt.year)

ccm = ccm.loc[~(ccm['fyear_from'] > ccm['fyear_to'])].reset_index(drop = True)

ccm['fyear_to'] = ccm['fyear_to'].fillna(2024)

ccm = ccm[['gvkey', 'lpermno', 'fyear_from', 'fyear_to']]

between = ccm.to_dict(orient = 'index')

in_between = []

for k, v in between.items():
    in_between.append([i for i in range(int(v['fyear_from']), int(v['fyear_to']) + 1)])

ccm['fyear'] = pd.Series(in_between)

ccm = ccm.explode('fyear')

ccm = ccm.drop(columns = ['fyear_from', 'fyear_to'])

ccm.to_csv(os.path.join(os.getcwd(), '1_data', 'compustat_crsp_link.tsv'),
           sep = '\t', index = False)
