==> 2_code/0_get_crsp_ibes_link.ipynb <==
import wrds, os
import numpy as np
import pandas as pd

os.chdir('..')

---

db = wrds.Connection()

df = db.get_table(library = 'wrdsapps', table = 'ibcrsphist')

df.to_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'crsp_ibes_link_raw.tsv'),
          sep = '\t', index = False)

df = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'crsp_ibes_link_raw.tsv'),
                 sep = '\t')

scores:
1. 8-digit historical CUSIP match
2. Historical ticker match + 6-digit CUSIPs and similar company names
3. Historical ticker match + plus 6-digit CUSIPs, but different company names
4. Historical ticker match + similar company names, but different 6-digit CUSIPs
5. Historical ticker match, but both 6-digit CUSIPs and company names are different
6. No matching, or a one-to-many match that cannot be resolved


also: https://wrds-www.wharton.upenn.edu/pages/wrds-research/applications/python-replications/linking-ibes-and-crsp-data-python/

df = df.loc[df['score'] == 1]

df['date_from'], df['date_to'] = pd.to_datetime(df['sdate']), pd.to_datetime(df['edate'])

df['fyear_from'] = np.where(df['date_from'].dt.month < 6,
                            df['date_from'].dt.year,
                            df['date_from'].dt.year + 1)

df['fyear_to'] = np.where(df['date_to'].dt.month < 6,
                          df['date_to'].dt.year - 1,
                          df['date_to'].dt.year)

df = df.loc[~(df['fyear_from'] > df['fyear_to'])].reset_index(drop = True)

df['fyear_to'] = df['fyear_to'].fillna(2024)

df

df = df.drop(columns = ['sdate', 'edate', 'date_from', 'date_to', 'score', 'ncusip'])

between = df.to_dict(orient = 'index')

in_between = []

for k, v in between.items():
    in_between.append([i for i in range(int(v['fyear_from']), int(v['fyear_to']) + 1)])

df['fyear'] = pd.Series(in_between)

df = df.explode('fyear')

df = df.drop(columns = ['fyear_from', 'fyear_to'])

df.to_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'crsp_ibes_link_final.tsv'),
          sep = '\t', index = False)
