==> 2_code/0_get_ibes_data.ipynb <==
import wrds, os
import numpy as np
import pandas as pd

os.chdir('..')

---

db = wrds.Connection()

# VU does not seem to have access anymore...

df = db.raw_sql("""select *
from ibes.statsum_epsus
where fiscalp = 'ANN'
and fpedats > '2005-01-01'
""")

df.to_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'ibes_summary_file_raw.tsv'),
          sep = '\t', index = False)
