==> 2_code/1_get_compustat_data.ipynb <==
import os, re, wrds
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm_notebook as tqdm

os.chdir('..')

---

db = wrds.Connection()

df = db.raw_sql("""select gvkey, fyear, consol, indfmt, datafmt, popsrc, curcd, costat, at, ib, ni, sale, prcc_f, ceq, oancf, txfed, txdfed, pidom, xidoc, spi, nopi, glp, sppiv, xrd, dvc, csho, txndb, txdba, txdbca, txndba, tlcf
from comp.funda
""")

df = df.loc[df[['gvkey', 'fyear']].notna().all(axis = 1)]

df = df.loc[(df['consol'] == 'C') & (df['indfmt'] == 'INDL') &
            (df['datafmt'] == 'STD') & (df['popsrc'] == 'D'),]

df = df.drop_duplicates(subset = ['gvkey', 'fyear'], keep = False)

df = df.sort_values(['gvkey', 'fyear'], ascending = True)

df.to_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'compustat.tsv'),
          sep = '\t',
          index = False)
