==> 2_code/2_add_database_identifiers_and_other_data.ipynb <==
import os, re
import numpy as np
import pandas as pd

os.chdir('..')

---

df = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'compustat.tsv'),
                 sep = '\t')

add historic header variables

comphist = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'comphist_prepared.tsv'),
                       sep = '\t')

df = df.merge(right = comphist, on = ['gvkey', 'fyear'], how = 'left')

add header variables and replace missing values on the historic variables by the header; then drop header variables

comp_header = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'compustat_header.tsv'),
                          sep = '\t')
comp_header = comp_header.drop(columns = 'loc')

df = df.merge(right = comp_header, on = 'gvkey', how = 'left')

for var in ['fic', 'sic', 'cik']:
    df[f'h{var}'] = df[f'h{var}'].fillna(df[var])

df = df.drop(columns = ['fic', 'sic', 'cik'])

add CRSP PERMNO (based on CCM)

ccm = pd.read_csv(os.path.join(os.getcwd(), '1_data', 'compustat_crsp_link.tsv'),
                  sep = '\t')

df = df.merge(right = ccm, on = ['gvkey', 'fyear'], how = 'left')

df = df.rename(columns = {'lpermno' : 'permno'})

add IBES ticker (based on CRSP-IBES link)

ibes = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'crsp_ibes_link_final.tsv'),
                   sep = '\t')

ibes = ibes.rename(columns = {'ticker' : 'ibes_ticker'})

df = df.merge(right = ibes, on = ['permno', 'fyear'], how = 'left')

load VA data and merge to the Compustat records; using the cik-fyear

va_info = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'va_data_info.tsv'),
                      sep = '\t')

va_info = va_info.drop_duplicates(subset = ['cik', 'fyear'])

va_info = va_info.rename(columns = {'cik' : 'hcik', 'sic' : 'sic_from_sec'})

df = df.merge(right = va_info, on = ['hcik', 'fyear'], how = 'left')

add valuation allowance data

va = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'va_data_clean.tsv'),
                 sep = '\t')

df = df.merge(right = va[['adsh', 'tag', 'va', 'zero_inferred']], on = 'adsh', how = 'left')

add DTA (gross)

dta = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'dta_gross.tsv'),
                  sep = '\t')

df = df.merge(right = dta, on = 'adsh', how = 'left')

add imputed TLCFs

imp = pd.read_csv(os.path.join(os.getcwd(), '1_data', 'imputation_upload_2024.txt'),
                  sep = '\t')

df = df.merge(right = imp, on = ['gvkey', 'fyear'], how = 'left')

add additional info from imputation paper

imp_add = pd.read_csv(os.path.join(os.getcwd(), '1_data', 'imputation_upload_2024_specifics_footnote_paper.txt'),
                  sep = '\t')

df = df.merge(right = imp_add[['gvkey', 'fyear', 'tlcf_pre_2018', 'tlcf_post_2018']], on = ['gvkey', 'fyear'], how = 'left')

save to disk

df.to_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'sample_raw.tsv'),
          sep = '\t', index = False)
