==> 2_code/2_prepare_ibes_data.ipynb <==
import os, re
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm_notebook as tqdm

os.chdir('..')

---

df = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'ibes_summary_file_raw.tsv'),
                 sep = '\t')

keep relevant columns

df = df[['ticker', 'fpedats', 'statpers', 'anndats_act', 'actual', 'meanest', 'medest', 'numest']]

delete:

1. if actual EPS is missing
2. if forecast data is after annual report date
3. forecast is not the latest forecast (i.e., only keep latest forecast)

df = df.loc[df['actual'].notna()]

df['statpers'] = pd.to_datetime(df['statpers'])
df['anndats_act'] = pd.to_datetime(df['anndats_act'])

df = df.loc[df['statpers'] == df.groupby(['ticker', 'fpedats'])['statpers'].transform('max')]

sort

df = df.sort_values(by = ['ticker', 'fpedats', 'statpers'])

get fyear

df['fpedats'] = pd.to_datetime(df['fpedats'])

df['fyear'] = np.where(df['fpedats'].dt.month < 6,
                       df['fpedats'].dt.year - 1,
                       df['fpedats'].dt.year)

save to disk

df.to_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'ibes_summary_file_clean.tsv'),
          sep = '\t', index = False)
