==> 2_code/3_get_crsp_data.ipynb <==
import os, re, wrds
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm_notebook as tqdm

os.chdir('..')

---

meta = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'sample_raw.tsv'),
                   sep = '\t')

meta = meta.loc[meta['permno'].notna()]

---

db = wrds.Connection()

permnos = [str(int(p)) for p in meta['permno'].unique()]

permnos = "', '".join(permnos)

test = db.get_table(table = 'dsf_v2', library = 'crsp', obs = 10)

db.list_tables('crsp')

crsp = db.raw_sql("""select permno, date, vol, numtrd, ret, retx, prc
from crsp_a_stock.dsf
where permno in ('{}')
and date > '2005-01-01'
""".format(permnos))

crsp.to_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'crsp.tsv'), sep = '\t', index = False)

---

db = wrds.Connection()

sp500 = db.get_table('crsp_a_indexes', 'dsp500')

sp500.to_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'sp500.tsv'), sep = '\t', index = False)
