==> 2_code/3_prepare_analysis.ipynb <==
import os, re
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm_notebook as tqdm
import matplotlib.pyplot as plt

os.chdir('..')

---

df = pd.read_csv(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'sample_raw.tsv'),
                 sep = '\t')

df = df.sort_values(by = ['gvkey', 'fyear'])

create 2 digit SIC groups

df['sic_group'] = np.floor(df['hsic'] / 100).astype('Int64')

create loss sequence variable

# first create variable to check whether years are consecutive
df['L1_fyear'] = df.groupby('gvkey')['fyear'].shift(1)
df['is_year_cons'] = np.where((df['fyear'] - df['L1_fyear']) == 1, 1, 0)

out = []
loss_count = 0

for i, r in df.iterrows():
    if r['is_year_cons'] == 0:
        loss_count = 0
    elif pd.isnull(r['ib']):
        loss_count = 0
    elif r['ib'] >= 0:
        loss_count = 0
    elif r['ib'] < 0:
        loss_count += 1
    
    # only append from second loss onwards (otherwise 'sequential' doesn't really make sense?)
    # and: there is also a firstloss dummy in the data that captures the first instance of a loss
    loss_count_to_append = 0 if loss_count == 1 else loss_count
    
    out.append(loss_count_to_append)

df['lossseq'] = out

# the way the code above is set up, when ib is missing, lossseq = 0;
# set those observations to NA because we don't know really what ib in these cases
# so we can't conclude it should be a loss sequence of zero
df['lossseq'] = np.where(df['ib'].isna(), np.nan, df['lossseq'])

df.to_stata(os.path.join(os.getcwd(), '3_pipeline', '1_intermediate', 'sample.dta'),
            write_index = False)

---
